auth/libraries/pip-auth/requirements.txt
pyjwt==1.7.1
requests==2.22.0
pytest==5.3.4
simplejson==3.17.2